Deleted wms_sources
