+++ server/rest/wms_dataset.py
@@ -25,16 +25,66 @@ from girder.api.rest import loadmodel
 from girder.api.rest import getUrlParts
 from girder.constants import AccessType
 
+from owslib.wms import WebMapService
+
 from girder.plugins.minerva.rest.dataset import Dataset
 from girder.plugins.minerva.utility.minerva_utility import decryptCredentials
+from girder.plugins.minerva.utility.minerva_utility import encryptCredentials
 
 
 class WmsDataset(Dataset):
 
     def __init__(self):
         self.resourceName = 'minerva_dataset_wms'
-        self.route('POST', (), self.createWmsDataset)
+        self.route('POST', (), self.createWmsSource)
 
+    @access.user
+    def createWmsSource(self, params):
+        def sourceMetadata(username, password, baseURL, hostName, name):
+            minerva_metadata = {
+                'source_type': 'wms',
+                'wms_params': {
+                    'base_url': baseURL,
+                    'host_name': hostName
+                },
+                'wms_source_name': name
+            }
+
+            if username and password:
+                credentials = encryptCredentials("{}:{}".format(
+                    username, password))
+                minerva_metadata['wms_params']['credentials'] = credentials
+
+            return minerva_metadata
+
+        name = params['name']
+        baseURL = params['baseURL']
+        parsedUrl = getUrlParts(baseURL)
+        hostName = parsedUrl.netloc
+        username = params['username'] if 'username' in params else None
+        password = params['password'] if 'password' in params else None
+        wms = WebMapService(baseURL, version='1.1.1',
+                            username=username,
+                            password=password)
+        layersType = list(wms.contents)
+        layers = []
+        source = sourceMetadata(username, password, baseURL, hostName, name)
+
+
+        for layerType in layersType:
+            layer = {
+                'layer_title': wms[layerType].title,
+                'layer_type': layerType
+            }
+
+            dataset = self.createWmsDataset({'meta': {'minerva': source}},
+                                                  params={'typeName': layer['layer_type'],
+                                                          'name': layer['layer_title']})
+
+            layers.append(dataset)
+
+        return layers
+       
     @access.user
     def createWmsDataset(self, wmsSource, params):
         baseURL = wmsSource['meta']['minerva']['wms_params']['base_url']
@@ -75,6 +125,7 @@ class WmsDataset(Dataset):
             minerva_metadata['credentials'] = credentials
         dataset = self.constructDataset(name, minerva_metadata)
         return dataset
+
     createWmsDataset.description = (
         Description('Create a WMS Dataset from a WMS Source.')
         .responseClass('Item')
